requirements.txt
Flask
requests
google-genai
gunicorn